09_minikube/example/project/payment/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
amqp==5.1.1
    # via kombu
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r requirements.in
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
grpcio==1.53.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.53.0
    # via -r requirements.in
kombu==5.2.4
    # via celery
mysql-connector-python==8.0.29
    # via -r requirements.in
prompt-toolkit==3.0.38
    # via click-repl
protobuf==4.22.1
    # via
    #   grpcio-tools
    #   mysql-connector-python
pytz==2023.2
    # via celery
redis==4.5.3
    # via celery
six==1.16.0
    # via click-repl
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
